requirements.txt
asgiref==3.8.1
Django==4.2.18
sqlparse==0.5.3
tzdata==2024.2

# Кэширование и очереди
redis==5.2.1
django-redis==5.4.0
celery==5.4.0
flower==2.0.1

# Мониторинг и метрики
prometheus-client==0.21.1
django-prometheus==2.3.1
python-json-logger==3.2.1

# API и безопасность
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
PyJWT==2.8.0
django-cors-headers==4.6.0
django-filter==24.3

# Редакторы и загрузка файлов
django-ckeditor==6.7.2
Pillow==11.1.0

# Дополнительные утилиты
python-dateutil==2.8.2
pytz==2023.3

# Тестирование
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
